Reworked the generic star maps available.
This removes the Dvaered Core, Za'lek Core, and Sirius Core star
maps from the main game (they still exist on Wakomi) and adds a new
"Dvaered-Za'lek trade route" map. The intent behind this is to give
the player more exploration to do through space, since exploration
is one of the big things the player is expected to enjoy doing in
this game and being given maps of the "cores" of various factions
sort of cheapens that. You can now find at least some of the main
factions' systems, but like with the Soromid, you have to explore
directly to properly discover them. (In truth previous maps only
covered half their territory at most, but this should make the
exploration feel more meaningful since you don't get the "main"
parts.)

Empire and Frontier are left available as singular maps for separate
reasons. In the case of the Empire map, it's because the Empire is the
first place you're likely to explore anyway. In the case of the Frontier,
it's because Frontier space is rather small and since they're
a small faction, I don't know yet if I want to make some sort of
border systems map to replace it (maybe I'll make an "FLF-Dvaered
Conflict Zone" map or something, I need to think about it some more).