Hardened credits2str so that it never creates bad strings with rounding.
The previous implementation would e.g. represent a value of
999999999 as "1000 M¢" due to accurately rounding up (something I
noticed while testing something else accidentally). This commit adds
code that takes that into account and ensures that such a case will
accurately round up to "1 G¢" rather than "1000 M¢". Had to look up
what sort of rounding is supposed to be used, and found that the type
of rounding used is "round to even":

https://stackoverflow.com/questions/10357192/printf-rounding-behavior-for-doubles