Adjusted the Llama a bit to hopefully make it better at the start.
The Llama is a bit weird because it starts with the smallest possible
engines, but a -20 turn penalty makes it less than ideal for those
and especially for forward-facing weapons. Made its stats slightly
more speed-oriented at a loss of cargo space and also added the
Gawain's special launcher penalties for good measure.