Fixed player.landWindow not updating tabs during the tab switch.
Noticed this while doing the DIY Nerds mission; it caused the slot
images to not be visible when that mission automatically kicked me
to the Equipment tab (along with some other missing details I didn't
pay attention to). With the addition of calling land_changeTab, this
problem is avoided and player.landWindow works as expected. This also
ensures that hooks get called normally.