Corrected update-po.sh implementation, update potfiles.
The script was relying on the existence of a binary called "python".
This simply does not exist on all systems; Debian 11 is one such
system which does *not* opt to provide a "python" binary by default.

Also updated the potfiles list, which was what I was originally
going to do before discovering that build script flaw.