Reorganized the Outfits tab to make the information more compact.
This moves the common information to the top (since it doesn't need
as much space as actual stats), removes the redundant "Owned" part,
merges the header portion with the contents for the common info,
and merges slot type and size into one line. In total this saves a
lot of space, and while I was at it, I made all of the text fields
dynamically sized so that if there's too much description text, it
will get cut off instead of overflowing onto the buttons.