Fixed and improved ship sorting.
The previous sorting method was based on the old class definitions
and was based on… the position of the class in the enum. That caused
completely arbitrary ordering. To correct this, I changed it to based
first on rarity, then price, then mass. Seems to be a good sorting
method.