Removed Sindbad "fleet" and "base" AI.
These are both archaic and unused, with the "base" AI literally being
based on an API that no longer exists as it did when it was created.
It seems neither was actually in use (the "base" AI was in use only
by the "Sindbad" fleet), so we'll just remove them.