Made it so Patrol misn paying faction presence must be above enemy presence.
This is to avoid generating missions that are frustrating to do
because of how outnumbered you are, and avoid weirdness of patrolling
enemy territory to eliminate enemies. This problem mainly only affected
the pirate patrol mission, but avoiding it with the regular patrol
mission makes sense too. It'll lead to patrol missions being less
common, but this drawback should ensure that all patrol missions are
reasonably doable.