Removed (almost) all references to "Sirians" and "Sirii".
I find "Sirian" as a term for citizens of House Sirius problematic
since it sounds exactly like "Syrian", a word the describes real-life
people (those from Syria). As such, all usages of it and "Sirii"
have been removed and replaced with "Siriusite" and "Siriusites"
(using the Greek suffix "-ite", which means "member of").

I didn't touch the Academy Hack campaign yet because I'm going to
do some more extensive work on that and want to do it all in one
go.