Changed the way Baron Prince spawns artifact hunters.
At first I was just going to fix the problem that they were depending
on a controlled attack command, but decided I really didn't like its
method of spawning the artifact hunters and modified it to now spawn
a bunch of *individual* artifact hunters at random times after jumping
in, with the exact number and ship types randomized. It is also able
to spawn more different types of artifact hunter ships.