Increased cargo penalty for Gawain and Schroedinger.
Someone on Naev's Discord mentioned using a Gawain as a cargo ship,
so I checked it (and the Schroedinger, also mentioned) to see if
that is actually an optimal setup since that would run counter to
the intended design of the game (and in particular make the Quicksilver
less compelling of a ship). I found that the Schroedinger is
sub-optimal, and while the Gawain is too, it's a very subtle disadvantage
compared to the Gawain assuming you're okay with sacrificing a little
cargo space for a faster speed, at least at a glance. I've therefore
changed the cargo mod for both ships to -50%, and also increased the
Schroedinger's base cargo to ensure it still has enough cargo space
for non-cargo missions that use small amounts of cargo space (when
equipped as intended). They are now almost equally (visibly) bad for
cargo, and this should hopefully dissuade attempting to use these
ships that way (and encourage using either a Llama or Quicksilver,
which are actually designed for this sort of job).