Improved the way ship stats are displayed on the Equipment tab.
This changes the color for "lost" stats to yellow, but also adds an
"out of range" conditional which, if true, leads to red coloring.
The yellow colored stuff is accompanied by an asterisk for colorblind
accessibility.

This additionally fixes a problem where speed stats were colored green
if zero, since code ensures they're never actually exactly zero.
Ultimately the solution I settled with is to consider <= 0.5 values out
of bounds, since the display doesn't show values below this other than
zero and such slow speeds are unbearable (in fact speeds about ten times
as much are unbearable). This leads to 0 speed values showing red,
which makes sense, because that usually means an engine hasn't been
equipped.