Add a cargo penalty to the Hyena as well.
The Hyena is quite a fast ship and so had the potential to seem like
a better candidate for fast cargo transport than the Quicksilver,
like the Gawain and Schroedinger, even though in reality the Quicksilver
was better. Built-in cargo capacity has been doubled so that in intended
builds, cargo capacity should be about the same with the -50% penalty.