Made representation of stats increase/decrease better and more accurate.
This commit also changes the behavior of the radar range stats so
that flat increases happen before multiplicative, rather than after,
bringing them in line with other similar stats. It has no real effects
on the game currently since flat increases weren't in use, but a good
change long-term I think.

As for how stats display has improved, when deciding whether a stat
has increased or decreased, it now checks against the ship's defaults
factoring in ship stats, rather than just the flat base stats. The
important benefit of this is that gains from outfits will be seen as
gains even if they don't make up for a loss the ship starts with due
to e.g. negative cargo_mod.