Added experimental smaller overlay in Legacy GUI, fixed overlay size stuff.
While adding experimental use of the thus unused gui.setMapOverlayBounds,
I noticed two problems with it: firstly, it was displaying the
rectangle behind contents incorrectly (using the right bound instead
of the bottom bound for the Y coordinate). Secondly, this wasn't reset
when a new GUI was loaded, which meant that GUIs either had to use
the function or allow others that used it to bleed their setting into
them. Both problems have been fixed and the experimental smaller overlay
in the Legacy GUI seems to work perfectly.