Cleaned up and simplified the Drunkard mission.
This gets rid of the hailing part of the pay message, for two reasons:
firstly, it can cause a problematic condition if the player doesn't
respond to the hail, which would have had to be handled by extra logic.
Second, it doesn't really add anything to the mission; it just sort
of splits the text up into smaller chunks. I didn't consider this when
writing the change, but it occurs to me now that the sequence has the
potential to have conflicts with other missions as well because no
system claims are made. The mission works just as well with a simple
takeoff hook and a lot more simply.

I also made it so that the mission doesn't check cargo space when
you accept the mission, since the cargo space of your *current* ship
doesn't really matter at that time; added a note of how many tonnes
you need to pick up to the OSD; and made it so that it doesn't fail
the mission if you land on the planet with too little space, and it
instead informs you that you don't have enough space (more specifically
than before) and tells you to add space and try again.