Increased potency of Trader virt assets, removed "independent" versions
The "independent" versions of trader assets only existed because of
the brief time when the Trader faction was split into the static
"Traders" faction and the normal "Traders Guild" faction. All traders
are static now, so there's no longer any need. However, I noticed
that the Independent version had more trader presence and so transcribed
that to the regular virtual trader presence assets.