Added a significant shield buff for the Koala.
The Koala is not very maneuverable (by design) and yet its small
size make it have a hard time fending off enemies through sheer
strength (and in fact that usually gets it killed). This is designed
to address that: the Koala now has a significant shield buff, which
should hopefully allow it to more effectively defend itself. When
combined with its fast speed (which enables outrunning things like
Destroyers easily), my hope is that this change will make the Koala
about as compelling as the Mule, but for different reasons.