Don't reset tc_rampdown when resetting speed.
The origin of this line seems to be c48c5f5, which explains its
reasoning as "we aren't ramping down anymore". I didn't have the
knowledge way back then to question this, but now I do, and this is
inaccurate. Setting tc_rampdown to zero forcibly like this actually
causes time to *speed up* at inappropriate times, which is the
source of sudden lurches in speed which have existed for a long
time now and shouldn't have existed.