New explanation for what Shadow is.
Thought about it a bit, and I'm going to make the Four Winds an
elaborate Proteron espionage scheme, with none of its agents aware
of the true purpose of the organization they are a part of. As a
result, the seemingly unscrupulous actions of other Four Winds
captains is actually to protect the galaxy from the Proteron: they're
trying to prevent the creation of new hypergates, while the leaders
of the Four Winds remaining after the Incident are trying their best
to get hypergates built in spite of political instability which would
normally render their creation nearly impossible.