Improved keyboard navigation on the map window.
A couple of changes:

Firstly, the "F" key used to open the "Find" window is documented
by the button now; it's the shortcut for the button.

Secondly, using the "autonav" key to activate autonav has been
removed, because the "A" key already serves that purpose and
attempting to use in-game controls within the GUI could lead to
conflicts which may make some controls not work correctly. As long
as GUI keyboard shortcuts aren't customizable, consistently following
that rule on the map window is safest.