Added a buffer for FLF standing to avoid accidentally becoming enemies.
Being enemies with the FLF is a massive game-changer, changing the
Frontier from a generally safe region to the most dangerous region
on the map. Having the starting player standing be 3 (like the Frontier)
instead of 0 means that an accidental stray bullet or something won't
cause the player to immediately be thrust into that situation.