Removed nebula18.
I am genuinely saddened by this. That nebula is *gorgeous*, and I
really wanted to keep using it. But it's just not feasible.

That nebula has compression artifacts which are basically invisible
normally. But since Naikari displays backgrounds at floating-point
positions, sometimes they can be warped around to make it look like
they're doing sub-pixel movements. Unfortunately, when this is done
with nebula18, its normally invisible compression artifacts suddenly
pop, creating what looks like a strange ripple effect that is very
noticeable.

I was able to get the ripple effect to disappear by rounding the
position of background images when drawn, but because backgrounds
move so slowly, this leads to an unacceptable jerkiness. I even
tried blurring the image a bit, but no dice. THe only way to stop
this problem is to remove the image.