Removed the annoying "Exit Game" option on the death menu.
The idea of giving the player the choice to either go to main menu or
quit game is nice *in theory*, but an "Exit Game" button that actually
fully quits the game on the game over screen essentially amounts to a
"rage quit" button, and it's practically asking to be a quit moment.
It's far, far more likely for someone to want to load the backup save
or a different save, and if someone does want to quit because they
died, an extra click or clicking the window "X" isn't a particularly
large extra step (and encouraging rage-quits seems like not a good
design choice).