Replace delayed claim hackiness for Shadow Vigil.
This causes all text to be handled in-mission, and to avoid annoyance
from a hail that gets you nothing, it adds special text when Shadow
Vigil is unable to make the needed claims giving a 10,000 credit
consolation payment (with the excuse that the pilot mistook you for
someone else). I chose to do this because the whole delayed claim thing
leaves open the possibility of the player being given a mission,
accepting it, and having it just not show up, which is a bad user
experience I think.