Slightly better escort event code (no functional change).
Was originally going to make escorts "eject" and return immediately
when you land, but it occurred to me that this feature would be in
conflict with the "docking" functionality as the only purpose of that
is to prevent ships from being destroyed while you fly a ship that
shouldn't be escorted by them. Need to think about this some more.
Maybe I'll want some sort of hybrid approach where a small fee is
assessed whenever an escort is destroyed. Or maybe I should keep
the system as-is. We'll see.