Display the full range of masses for launchers and bays.
Launchers and fighter bays have a mass that varies depending on how
full they are, which is something that ought to be noted. This can
be especially helpful for deciding whether or not a fighter bay should
go on a ship, since depending on play style you might have all the
fighters deployed all the time anyway (that's how a Mule would likely
use them for example).