Cleaned up es00 and improved it by making it not lose FLF rep.
The cleanup is just the standard style updates (including using the
fmt module and simplifying the OSD). In addition (and this was the
initial reason for starting this commit), the FLF ships spawned by
this mission have been changed from actual FLF ships to ships of an
FLF-derived dynamic faction so that the player firing on them doesn't
lead to reputation loss. This is primarily in case the player made
the bad decision to fly over and do the prisoner exchange with a
Destroyer or something, since this would leave the player blindsided
and pretty much force them to kill FLF ships (unlike in the VIP
mission where you know there's fighting breaking out from the start).
That's kind of unfair seeing how dangerous Frontier space becomes if
you make yourself an enemy of the FLF.