Overhauled handling of composite ship stats displays.
Previously the same string was used for both stat changes and overall
stats of a ship composited together. This change causes those to
be more thoroughly separated: instead of a "ship redundant" argument
specifying whether or not to display the stats for an overall ship,
there is now a separate ship for display of overall stats on a ship.
For most displayed stats, it is the same except using %.0f instead of
%G to avoid weirdness like printing "66.666667%" when "67%" is close
enough (miraculously hasn't happened with me at all, but best to
play it safe). For absolute percentage stats, though (the main reason
for this change), this changes the "pp" display to "%", so you would
get. e.g. "97% Nebula Resistance (Armor)" instead of
"+97 pp Nebula Resistance (Armor)" for the ship's display. (While the
latter is technically accurate, it doesn't make much sense as a way
to display the stat when it's the overall stat, not just a modifier
of something.)

This also gets rid of legacy double-underscores in shipstats.c which
were originally there to do one of my least favorite things in
programming, attempt to make the stats into a giant aligned table.
We're not doing that anymore, so no sense keeping that naming;
D__ELEM has been renamed to D_ELEM, A__ELEM to A_ELEM, etc.