Fixed cinematics mode not aborting autonav if player was controlled.
This was a side-effect of code intended to prevent the player from
stopping auto-cooling initiated by a script. However, the
player.cinematics() function is script-initiated, not player-initiated.
Therefore, this commit adds a new "force" option to player_autonavAbort
which causes autonav to be aborted even if the player is under manual
control. This new "force" option is also used on the player.autonavAbort
function, since that too would be script initiated. I don't know if
there are any real applications to this, but if it's true that
autonavAbort stops auto-cooling, that could be a legitimate use
while the player is under manual control.