Change custom tags to use <b> instead of <em>, and add a colon.
Using <em> in those situations is not correct HTML usage. <em> has
a defined specific meaning, and it's not meant for use on headers.
Replaced with <b>, which is a markup for bold text that is there only
for style purposes and without indicating extra importance. Since
in this case it's just being used as a stylized "miniature header"
and a colon is what serves the meaning purpose, <b> therefore makes
the most sense. (<i> doesn't make as much sense, I don't think,
because using italics for header-type text isn't commonplace.)