Improved info window handling of manual control and some other stuff.
Main purpose of this was to get rid of the thing where the Info window
was completely inaccessible if the player pilot was under manual
control. This changes that: it now only stops you from aborting
missions and jettisoning cargo, since aborting missions can prevent
a mission from clearing the player pilot control state (the only
reason I can think of to block access to the Info window). Of course
it would be possible to pass on that possibility to missions through
their abort functions, but I can see an argument that doing this
correctly adds a lot of complexity to any mission controlling the
player pilot for very little gain (usually just having to wait for a
cutscene to finish). I'll evaluate over time whether or not to
remove this restriction, but for now at least it's less restricted
than before.

While I was doing this, I also found a couple other issues and fixed
them. Firstly, I fixed a problem where aborting missions with cargo
would leave the cargo list as-is, causing it to be inaccurate.
Secondly, I removed the restriction disallowing jettisoning while
landed and instead made it so that jettisoning while landed doesn't
produce a jettisoned container in space (but still removes the cargo).