Improved Animal Trouble, hardend race1/2 against abort during control().
As I was adding hardening to race1 and race2 against aborts while
the player was being manually controlled, I noticed that both the
Za'lek Test mission and Animal Trouble event were problematic because
they didn't make system claims, which meant they could come into
conflict with something like the Sirius Preach event and lead to two
control events, causing one of them to end sooner than expected.
I haven't fixed Za'lek Test yet (that's going to require a much bigger
fix, I think with an event that runs if you're using the test engine),
but I've fixed the Animal Trouble mission with several changes.

Animal Trouble now begins by claiming the system and immediately
disabling landing and jumping. Then, 5 seconds later, it starts the
previous event with a modified start message noting the landing and
jumping restriction. The event also clears the system and disables
spawning to avoid frustration from getting killed by a pirate or
whatever, and keeps cinematics mode on (I assume it was turning it
off before because there wasn't an option to keep the GUI on and it
really only wanted cinematics for the autonav abort).