Removed the "class name" API.
While the one thing it was being used for is potentially good, it's
implemented as a series of hardcoded definitions for specific class
names, a leftover from when planet classes were stiffly defined to
only ever be allowed to be those particular class names. If class
names that are more understandable are desirable, I think making the
class names themselves those more descriptive "names" would make a
lot more sense.