Made star / dust speed independent of brightness.
This involved a lot of banging my head against the wall that is useless
GLSL error messages, but finally got it to work. This makes it so
that relative star / dust speed is independent of brightness, rather
than determined by it, allowing dim dust particles to still be fast.

This also removes a rather strange formula from stars.vert that
could depending on relspeed (previously brightness) lead to division
by zero or *negative* movement, and uses relspeed directly (the weird
formula seems to have been meant to expand a [0,0.8] value range to
a (0.1,1] value range, which I can only assume has to do with a
desire to use the brightness values a certain way or something).

Taking the opportunity, this commit also makes dust much, much less
bright.