Added a CPU bonus and speed malus to the Hephaestus.
The Hephaestus is a massive beast and was barely capable of even holding
the number of Za'lek drone bays that it is physically capable of taking.
This CPU bonus helps it substantially, balanced out by a substantial
speed malus. The speed malus also helps distinguish it better from
the Diablo, which is able to be much faster due to its low mass.