Replaced the outfit type indicator with an outfit size indicator.
This shows "S", "M", and "L", indicating size, instead of showing
the first character of the outfit type's English name. This method,
incidentally, is better than what was used for the previous function,
in that it can actually be translated. I could see Japanese using
something like 小, 中, and 大 for the size character, for example.