Equalized fighter bay amounts.
This makes all full-sized fighter bays have 3 capacity, and makes
all mini bays have 2 capacity, except for the Za'lek bays which have
4 capacity for fighters and 3 for bombers. The reason for this change
is because the previous setup, while making sense on paper (choosing
between fewer strong ships and more weak ships), the fact that the
stronger bays were heavier and the fact that outnumbering opponents
is usually better than relying on more firepower meant that the light
bays were obviously better. This changes the dynamic: heavier fighters
are better, but with the caveat that they take up more mass.

The Za'lek Bomber Drone Fighter Bay is an exception: bomber drones have
a specific advantage of being long-range, and they're not much
heavier than heavy drones, so it makes sense to limit their capacity,
and this shouldn't make them any less desirable. It might even be
used in combination with other Za'lek drone bays.