Replaced system map by double-click with "System Info" button.
Based on 914c446 by @jjustra, but with some changes, mostly for Naikari
compatibility and avoiding merge conflicts. In addition to that and
style changes, this not only adds the button, but also removes the
old double-click method of opening the system info screen. The reason
for this is I've always found this "double click to open" control
rather annoying, because it's not actually a double-click, it's a
single click if a system is selected, which has led me to accidentally
open the "System Map" many times when I was just trying to select a
system that I didn't notice was already selected. Now that the button
is there, there's no need for causing that annoyance.