Cleaned up sh02_tapping and changed its story.
The whole thing about House Sirius selling ships to the Frontier and
Nexus caring in the slightest just doesn't make much sense, so I've
replaced that with a new story about an upstart, like Krain Industries,
cropping up in Sirius territory. I'll need to change the rest of the
missions to match up with this, of course.