Allow misn.markerRm to accept nil, and in case of nil do nothing.
This is helpful because it avoids the need for if-guards in the Lua
code of the form `if marker ~= nil then misn.markerRm(marker) end`.
The code already accepts already removed markers, so it might as
well accept markers that don't exist yet and simplify the Lua code
a bit.