Removed if guards around hook.rm.
These were never necessary, but I tended to put them there because
I couldn't necessarily remember if it was hook.rm or misn.markerRm
that required them (it was the latter). Now that I've fixed the
other one, best to get rid of these if guards (or at least the ones
I can immediately see).