Got rid of the rest of the Shark campaign.
I really don't feel like the rest of the Shark campaign that exists
is particularly worth keeping and fixing up. A lot would have to be
changed, and they would ultimately just be simple and uninteresting
missions. I think it would be best to leave it off at Tapping for
now, and maybe consider expanding the campaign later on if other
interesting ideas (possibly intermingling with the story as the
campaign used to intermingle with the Collective campaign) spring
to mind. For now, cutting it short I think leaves it sufficiently
satisfying for what it is.