Changed "MN/t" to "km/s²", and "Thrust" display to "Acceleration".
"MN/t" has the exact same meaning as "km/s²", but is needlessly
technical and makes what is being indicated less clear. These displayed
values are in fact acceleration values and *not* the internal thrust
(force) values used by the engine, so I've also made them display as
"Acceleration" instead of "Thrust", to be more clear what this stat is.