Added my own more aggressive fix to division by zero in ai.face().
This removes the undocumented feature to pass -1 to cause a pilot to
face itself (which doesn't make sense and would have always caused
division by zero for the past decade), and it causes a pilot to not
attempt to face the target at all if its position and the target's
are exactly the same (since no direction is objectively right in that
case). Noted in a comment that this method could be slightly improved
by considering the velocities of the pilot and target and predicting
the aim angle based on that, but that would be a single-frame
improvement and I don't have the spoons to work on that right now.