Cleaned up achack01, disabled achack02, removed achack03 and 04.
I'm satisfied with the idea of this campaign, but not with the
follow-ups, which to me feel rushed and not very well thought-out.
So given that, I've removed 3 and 4, disabled 2 (didn't remove it
because I want to still be able to reference its telling of the
campaign's story background), and fixed up and enhanced achack01.
In addition to style and convention fixes, I also made the player
question the fact that a Fyrra-class Siriusite wants to kill a
Serra-class Siriusite right from the beginning, rather than having
the player act surprised only on confirmation that they're military
(which is something you would expect of a Serra-class Siriusite on
a planet like this). I also got rid of a claim whose only purpose
was to prevent minor weirdness with another mission using the same
system (it was specifically for Dark Shadow), which is a misuse of
the claim system and wholly unnecessary.