Got rid of the music delay feature, fixing problems.
There has been a feature in Naev for years which forces music to
play for 10… seconds, I think. Trawling through history is a nightmare,
but from what I can gleam, the only purpose of this feature is to
force combat music to play for longer than it's supposed to. I was
running into a case where this feature was forcing combat music to
continue playing in an entirely empty system. This was caused by the
fix to intro music continuing to play in the system when starting a
new game: what that did was set the "last" music to "ambient", which
sounds good on paper, except the check right above it disables
forcing of ambient music if the last music was ambient (to avoid
ambient music overriding other ambient music).

In summary, this means that I had two choices: either return to the
old status quo where intro music can bleed into the game, or remove
the music delay feature. I chose the latter. I honestly don't see
what benefit there really is from forcing combat music to continue
playing after the threat is over just because the threat ended
quickly.