Fixed a problem with music chooser disabled check, removed NULL choice use.
There was a mistake in how disabled hostiles were handled, in that
both parts that updated player.disabled_enemies relied entirely on
pilot_isHostile() which is an incomplete check. The PILOT_HOSTILE
flag also had to be checked, since pilot_isHostile() returns true
for all enemies, not just those that are actively hostile.

This also removes the use of NULL in one of the music_choose calls.
The Lua code makes it sound like the meaning was vague, but that
call was when a new system was being initialized, and since in a
new system you by default won't have hostiles, that means you always
want to choose ambient, and indeed the Lua script was just treating
nil as synonymous to "ambient".