Allow nil values in certain fleet.add tables, sameify nil spawn point.
There is some code which I wrote at some point where I made a mistake
and wrote nil as pilot names in a table, but the problem with that
is that nil values cannot exist within an ordinal Lua table. However,
it occurred to me that the only table that gets indexed ordinally
rather than by explicit key is the ships table; all the others are
indexed with the assumption that they are the same length as the
ship table and that each index refers to the argument to pass to the
given ship. So for those parameters that support nil in pilot.add(),
I've codified that they're allowed to be nil when passing a table,
too.

This also adds a feature where a nil non-table location leads to
spawning all ships in the fleet from the same randomly chosen
location, which should be helpful for a lot of cases.