Added direct recognition of empty titles for dialogs used by Lua.
This causes those dialogs to make use of the space that would normally
be used for the title in the case that the title is a completely empty
string. This avoids a weird effect where windows otherwise have more
space at the top than they actually need, leading to a cleaner look.