Fixed some Shark mission problems.
This in particular changes the first mission to kill a Vendetta
rather than an Ancestor, since I found that killing an Ancestor
with a Shark is now much harder than it used to be (due to a combination
of the AI piloting bombers better and bombers being more competitive).
The Vendetta is a much more reasonable ship to fight against as a
Shark because while it can kill you quite quickly, its shots are
easier to dodge around without spending long periods of time trying
to survive a bombardment of missiles.