Added a Test Engine event, plus fixed a couple things.
The couple things I fixed are deleting the achack04 event (no longer
used) and fixing player.autonavAbort to make the string argument
optional.

The Test Engine event is a replacement for the existing Za'lek Test
Engine mission's problems functioning. That functionality has a
couple of problems: that it doesn't work outside of the mission, and
more importantly, that it can collide with other missions and events
because no claims are made for it. This event fixes both problems.

I will replace the Za'lek Test mission with a simpler replacement
later. I'll also be improving the Za'lek Test Engine and making it
actually viable… with the exception of the random problems it causes.
I might even consider splitting it into three separate engines, though
I probably won't go quite that far.