Improved the Pinnacle and Baron's prowling.
There was a common issue where the Pinnacle would sort of orbit one
of the points that it was trying to go to. Fixed that by limiting
the Pinnacle's speed to a very small number.

Also improved how the Pinnacle behaves slightly overall. Of note:
fixed some discrepancies with the first Baron mission, made the Prowl
event use claims rather than a list of manual checks (and both
Baron missions use claims as well), and made the Pinnacle explicitly
land rather than giving it AI control when the missions end.