Isolated viewport shenanigans so they don't cause unintended side-effects.
Viewport fudging intended just for the toolkit stuff was bleeding over
into the next frame's input, causing clicks on the edge radar to be offset
if a custom viewport is used. This fixes that by isolating the
viewport shenanigans to the GUI rendering and input handling, immediately
restoring the default viewport afterward.

This required inserting the same shenanigans to the place where TK
input is handled, and it also required resetting the Slim GUI's fps_y
setting so it behaves as before.