Music experiment: don't use land or takeoff music.
Just an idea I had to make music a little better: this change causes
land and takeoff music to be the same as ambient music, and modifies
ambient music to handle the player being landed based on the landed
planet's faction. This avoids constantly repeating those land and
takeoff tracks and also connects themes better, or at least that's
the idea.

Of course, what I ultimately want to do is replace the entire
soundtrack with that big ambitious music project, but that's so big
and ambitious it's probably years down the line. So let's give this
a try and see how it goes.