Adjusted slot positioning for consistency, made Kestrel fighter slots medium.
This commit is primarily designed to make it so that the slots are
intentionally positioned such that placing the same kinds of weapons
in the same position will lead to results that make sense. While I
was doing this, I decided to also change the Kestrel from having two
large fighter bay mounts to having two medium fighter bay mounts, on
the grounds that the Kestrel is extremely light, barely  heavier than
a Destroyer, and so should have that small limitation that they have
rather than being the same as their larger cousins.

Also fixed a mistake I saw in the Laser Turret MK1's description.