Added a default weapon configuration to all ships.
Wow, what an effort. This was tedious, but I think worth it. With this,
any ship you buy now starts out usable even if not at its full
potential, which is much more beginner-friendly than tasking them
with figuring out the best weapons to use right away.

In tandem with this, I've done some related things:
* Updated the tutorial's references to suggest that ships normally
  come with weapons by default.
* Removed start script's equipping of laser cannons.
* To enable the start script change, fixed a bug where new players
  would not have their weapons automatically handled until something
  in particular called for it, rendering a ship's default weapons
  temporarily unusable.