Added simple AI logic to always launch fighter bay fighters.
This logic is simple, but it should be perfectly fine for the most
part. It ensures that all AI players will now always launch all of
their fighters, rather than ignoring fighter bays and being weak and
vulnerable as a result. This should make them an incredibly deadly
force.

It can be argued that a more complex behavior would be better, but
in truth there is rarely a good reason to hold fighters back. The
only complex behavior that would improve this is to consider engaging
active cooldown because it's run out of fighters.