Added a "Current Objective" display to Info window mission tab.
This simply shows the currently highlighted OSD objective. This is
useful e.g. if you have a whole lot of missions or if you need to
know your current objective while you're looking at the info window.

This commit also fixes a regression in the last commit where the
rewards display was a restricted size.