Made centered widget text fall back to uncentered if too wide.
Centered text is always one line only. To compensate for this, added
logic that only uses centered printing if there's enough width for
all of the text. This allows centered text to still have the benefit
of being allowed to wrap text if it's too long, although it still
does not permit newlines in centered text.