Replaced Improved Refrigeration Cycle with Rotary Turbo Modulator.
The Improved Refrigeration Cycle was easily one of the worst outfits
since it was essentially just a worse, small-size version of the
Targeting Array outfit. That meant it could only theoretically be
useful for small couriers, which typically don't have to worry about
heat and wouldn't be able to last enough battles for heat to matter.
This new outfit is much more useful to them (providing a fire rate
bonus, which in turn means more damage), and also potentially useful
for larger ships as well.